Move the worker files outside of the example
