+++ src/worker/AnalysisWorker.js
