+++ src/worker/analysis.worker.js
